Share more code for name-prefix handling.
Namecoin and Emercoin both have special treatment for name prefixes on
scripts.  Through the NameMixin, they were already sharing some code for
that.  There was, however, still a lot of duplicated logic.

This refactors the code so that more logic is shared in the common
NameMixin.  In particular, there is now a general function for matching
a potential name prefix against the allowed structures of name operations.
Thus, the coin-specific code only has to define the valid structures of
name operations in the coin and handle coin-specific extra work (like
building the normalised script that is used for indexing names in
Namecoin).