Split out deserialisation of CAuxPow.
The current DeserializerAuxPow has code for reading a block header
including (potentially) an auxpow.  In this change, we split out the
logic for reading the actual CAuxPow instance to its own function.

This makes the code easier to read, as the different steps are more
clearly separated.  It will also be useful for working with CAuxPow
instances that are part of a different serialisation format (e.g.
for the Xaya blockchain).