Add Xaya coin with deserialiser.
This defines a new coin for Xaya (https://github.com/xaya/xaya), with
the appropriate deserialiser methods.  Xaya is based on Namecoin, but
uses its own special twist on auxpow for the PoW, so we need to handle
that properly in block-header deserialisation.

Xaya also has names like Namecoin (but with slightly changed semantics),
but handling for them (i.e. building a name index in addition to the
address index) will be added later on.