Create NameIndexMixin from Namecoin.
This factors out some code from Namecoin to a new NameIndexMixin.  This
code is what we need to keep track of a name index in addition to the
address index, as done by the NameIndexBlockProcessor.

With the code factored out, the Coin itself only needs to define the
set of possible name operations it allows (and where they have the name
by which the index will be built).

For now, this is only used by Namecoin.  But in the future, we will use
it for Xaya as well.