Add name index for Xaya (#890)
* Rename NamecoinBlockProcessor.

For Xaya, we will use the same kind of logic for keeping track of an
index of names in addition to the index of addresses.  Thus it makes
sense to call it by some more generic name.  Here, we rename it to
NameIndexBlockProcessor.

* Create NameIndexMixin from Namecoin.

This factors out some code from Namecoin to a new NameIndexMixin.  This
code is what we need to keep track of a name index in addition to the
address index, as done by the NameIndexBlockProcessor.

With the code factored out, the Coin itself only needs to define the
set of possible name operations it allows (and where they have the name
by which the index will be built).

For now, this is only used by Namecoin.  But in the future, we will use
it for Xaya as well.

* Add name indexing to Xaya.

Change the Xaya definition so that it handles name scripts correctly:
We want to strip them off the "real" address for indexing by address,
and we want to keep a separate index on names like Namecoin does.